+++ src/main/java/com/app/shahbaztrades/model/dto/order/TradeOrderResponse.java
@@ -5,12 +5,14 @@ import lombok.Builder;
 import lombok.Data;
 import lombok.experimental.FieldDefaults;
 
+import java.math.BigDecimal;
+
 @Data
 @Builder
 @FieldDefaults(level = AccessLevel.PRIVATE)
 public class TradeOrderResponse {
     String orderId;
     String status;
-    double averagePrice;
+    BigDecimal averagePrice;
     int pendingQuantity;
 }
